fix(filter-field): Fixes an issue with the free-text submission.
Quickly input free text submissions were ignored if the user was submitting the filter under the
defined debounceTime as we were only updating the value after the debounceTime.
We are now updating the value immediately, but we are handling the submission and necessary updates
to the component after the debounceTime expired.

Fixes #726